Renamed ingress ClusterOperator name from 'openshift-ingress-operator' to 'ingress'
