+++ pkg/operator/controller/status_test.go
@@ -1,4 +1,4 @@
-package clusteringress
+package controller
 
 import (
 	"fmt"
